Prof Baeta

+++ src/services/post.js
@@ -30,9 +30,21 @@ const post = await getPostById(id);
 return post;
 };
 
-const deletePost = async (id) => {
+const oi = async (id) => {
+const post = await BlogPost.findOne({ where: { id }, include: { model: User, as: 'user' } });
+return post;
+};
+
+const deletePost = async (id, user) => {
+  const post = await oi(id);
+  if (!post) return { type: 'oi', message: 'Post does not exist' };
+  console.log('postUserDataValuesEmail', post.user.dataValues.email);
+  console.log('user', user.username);
+  if ((post.user.dataValues.email !== user.username)) { 
+    return { type: 'tchau', message: 'Unauthorized user' }; 
+  }
  await BlogPost.destroy({ where: { id } });
- return true;
+ return { type: null, message: '' };
 };
 
 const searchPostByParams = async (q) => {
